Security, etc.

+++ WhatShouldIDoNow/DataAccess/TaskCommands.cs
@@ -22,8 +22,8 @@ namespace WhatShouldIDoNow.DataAccess
             {
                 var sql = @"
                     insert into TasksToDo
-                    ([Description], IntervalByHour)
-                    values (@Description, @IntervalByHour)
+                    ([Description], IntervalByHour, UserID)
+                    values (@Description, @IntervalByHour, @UserId)
                     select SCOPE_IDENTITY()";
 
                 var id = conn.ExecuteScalar<int>(sql, task);
@@ -31,40 +31,41 @@ namespace WhatShouldIDoNow.DataAccess
             }
         }
 
-        public TaskToDo GetRandomTaskWithDateStart()
+        public TaskToDo GetRandomTaskWithDateStart(int userId)
         {
             //Get a random task that started in the past
             string sql;
             sql = @"SELECT TOP 1 [Id],[DateCreated],[Description],[Category],[DateDue],[LastViewed],
                     [DateStart],[TimesViewed],[IntervalByHour] FROM [dbo].[TasksToDo]
-                    Where getdate() > [DateStart]
+                    Where getdate() > [DateStart] and UserID = @userId
                     Order by NEWID()";
             using (IDbConnection conn = _dbConnectionProvider.GetOpenWsidnConnection())
             {
-                var task = conn.QuerySingleOrDefault<TaskToDo>(sql);
+                var task = conn.QuerySingleOrDefault<TaskToDo>(sql, new { userId = userId });
                 return task;
 
             }
         }
 
-        public TaskToDo GetTask(int id)
+        public TaskToDo GetTask(int id, int userId)
         {
             string sql;
-            sql = "SELECT TOP 1 [Id],[DateCreated],[Description],[Category],[DateDue],[LastViewed],[DateStart],[TimesViewed],[IntervalByHour] FROM [dbo].[TasksToDo] where Id = @Id";
+            sql = "SELECT TOP 1 [Id],[DateCreated],[Description],[Category],[DateDue],[LastViewed]," +
+                "[DateStart],[TimesViewed],[IntervalByHour] FROM [dbo].[TasksToDo] where Id = @Id and UserID = @userId";
             using (IDbConnection conn = _dbConnectionProvider.GetOpenWsidnConnection())
             {
-                var task = conn.QuerySingleOrDefault<TaskToDo>(sql, new { Id = id });
+                var task = conn.QuerySingleOrDefault<TaskToDo>(sql, new { Id = id, userId = userId });
                 return task;
             }
         }
 
-        public void UpdateTaskDateStart(int id, DateTime dateStart)
+        public void UpdateTaskDateStart(int id, DateTime dateStart, int userId)
         {
             string sql;
-            sql = "UPDATE[dbo].[TasksToDo] Set DateStart = @dateStart Where ID = @id";
+            sql = "UPDATE[dbo].[TasksToDo] Set DateStart = @dateStart Where ID = @id and UserID = @userId";
             using (IDbConnection conn = _dbConnectionProvider.GetOpenWsidnConnection())
             {
-                conn.Execute(sql, new { id, dateStart });
+                conn.Execute(sql, new { id, dateStart, userId });
             }
         }
 
@@ -72,29 +73,29 @@ namespace WhatShouldIDoNow.DataAccess
         {
             var sql = @"
                 insert into TasksCompleted
-                ([Description], DateCreated, Category)
-                values (@Description, @DateCreated, @Category)";
+                ([Description], DateCreated, Category, UserID)
+                values (@Description, @DateCreated, @Category, @UserId)";
             using (var conn = _dbConnectionProvider.GetOpenWsidnConnection())
             {
                 conn.Execute(sql, task);
             }
         }
 
-        public void DeleteTaskTodo(int id)
+        public void DeleteTaskTodo(int id, int userId)
         {
-            var sql = "delete from TasksTodo where Id = @Id";
+            var sql = "delete from TasksTodo where Id = @Id and userId = @userId";
             using (var conn = _dbConnectionProvider.GetOpenWsidnConnection())
             {
-                conn.Execute(sql, new { Id = id });
+                conn.Execute(sql, new { Id = id, userId = userId });
             }
         }
 
-        public List<CompletedTask> GetCompletedTasks()
+        public List<CompletedTask> GetCompletedTasks(int userId)
         {
-            var sql = "SELECT [Id],[DateCreated],Description,[Category],[DateCompleted] FROM [dbo].[TasksCompleted]";
+            var sql = "SELECT [Id],[DateCreated],Description,[Category],[DateCompleted] FROM [dbo].[TasksCompleted] Where UserID = @userId";
             using (var conn = _dbConnectionProvider.GetOpenWsidnConnection())
             {
-                var results = conn.Query<CompletedTask>(sql);
+                var results = conn.Query<CompletedTask>(sql, new { userId = userId });
                 return results.ToList();
             }
         }
